Move stateful mixin into state.py
